Directory restructure - move into ./src
