+++ src/admin-client/package.json
