+++ src/admin-client/public/favicon.ico
