+++ src/admin-client/public/img/icons/android-chrome-maskable-192x192.png
