+++ src/admin-client/public/img/icons/apple-touch-icon-180x180.png
