+++ src/admin-client/public/img/icons/apple-touch-icon.png
