+++ src/admin-client/public/img/icons/favicon-32x32.png
