+++ src/admin-client/src/App.vue
