+++ src/admin-client/src/assets/logo.png
