+++ src/admin-client/src/components/EmojiButton.vue
