+++ src/db/models/poll.js
